fix: recreate UserAccount when deleted (on purpose or by accident)
When ensuring the UserAccounts for the MasterUserRecord exist, if one
is being deleted, then the controller will requeue (with a delay),
which should give time to the member cluster to finish deleting
the resources. Then, during the new reconcile loop, the
MasterUserRecord controller will recreate the UserAccount.

Also, small refactorings (func name and using `metav1.Now()`
as a shorthand)

Fixes https://issues.redhat.com/browse/CRT-778

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>