feat: add metric for returning users
Introduce a new type of metric for counter vectors
Use this new type to collect metrics for returning users.
When the user signs up for the first time, the `UserSignup`
resource (created by the Registration Service) gets a new
`toolchain.dev.openshift.com/activation-counter` annotation
with a value set to "1" (it has to be a `string`).
Then, at each new reactivation, the annotation value is increased
by one.

The `toolchain.dev.openshift.com/activation-counter` annotation does
not exist for UserSignup resources created prior to this PR,
so its value is set to "1" during the controller startup, when this latter
processes the existing resources.
Later when the user reactivates her account, the value is immediately set to "2".

Also, a new Prometheus `GaugeVec` metric called
`sandbox_users_per_activations` with an `activations` label
is exposed on the `/metrics` endpoint.

Note that this metric is NOT cumulative. For example, if we have 4 users:
- user 1 signed up 1 time
- user 2 signed up 2 times
- user 3 signed up 3 times
- user 4 signed up 4 times
then the metric will expose the following value:
```
sandbox_users_per_activations{activations="1"} 1
sandbox_users_per_activations{activations="2"} 1
sandbox_users_per_activations{activations="3"} 1
sandbox_users_per_activations{activations="4"} 1
```
in other words:
- 1 user signed up exactly 1 time
- 1 user signed up exactly 2 times
- 1 user signed up exactly 3 times
- 1 user signed up exactly 4 times

Also, since the metric also needs to be updated when the UserSignup
resource is deleted (manually or by the UserSignupCleanup controller),
the UserSignup controller now watches `delete` events on the UserSignup
resources, removes the finalizer and updates the metrics.

Also, a bit a refactoring on the init/sync funcs of the counters (with
comments)

Fixes https://issues.redhat.com/browse/CRT-988

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>