feat: metric to track phone verification failures
when a UserSignup resource is deleted because
the external user failed to pass the phone verification,
the cleanup controller checks if the user initiated an
attempt (and maybe some technical issue occurred)
or not (maybe the user did not want to input her phone number,
for privacy reasons).
This info is found by checking if the toolchain.dev.openshift.com/verification-code`
annotation has been set or not.

Fixes https://issues.redhat.com/browse/CRT-1064

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>