feat: reuse already existing tier templates in other tiers (#490)
## based_on_tier.yaml

This commit enables (with the help of a  new config file) reusing of already existing tier templates for defining a new tier. 
The idea is:
* everything that should be changed in the new tier is moved to parameters (to make it easily configurable) - you can see that in the changed `base` templates
* The new tier is defined via a `based-on-tier.yaml` file which says:
    * which tier should be reused
    * what the target tier is
    * the new values for the parameters
 for example:
```
from:
  name: base
to:
  name: baseextended
  parameters:
  - name: DEACTIVATION_TIMEOUT_DAYS
    value: 180 
```
which defines that for tier `baseextended` the logic should reuse the templates from `base` tier and replace the `DEACTIVATION_TIMEOUT_DAYS` parameter value with the new one: `180`

I believe that in this way, we can cover 90% of all use-cases and thus avoid code duplication as much as possible. 
I already configured `basedeactivationdisabled`, `basedeactivationdisabled`, `baseextended`, and `advanced`.

## Revisions

There is a small (I believe that cosmetic) change in the revisions that are used for the `TierTemplates`. Right now we take the revisions for each of the templates and then we use them in the name of the `TierTemplates`. However, most of the tiers won't contain the template files, but the `based_on_tier.yaml` file. And we need to catch the changes in both of them... 
So, this PR combines the revisions from both files - from the `based_on_tier.yaml` & from the particular template file that is being reused. If the tier doesn't contain `based_on_tier.yaml` file, then it just duplicates the same revision as is set for the template file.

let's consider this scenario:
```yaml
advanced/based_on_tier: "1234567"
base/cluster: "abc1234"
base/ns_dev: "bcd2345"
base/ns_stage: "cde3456"
```
as a result, there will be `TierTemplate` CRs with names:
* `advanced-clusterresources-1234567-abc1234`  
( `1234567` is revision of `based_on_tier.yaml` file, `abc1234` is revision of `base/cluster.yaml`)
* `advanced-dev-1234567-bcd2345`   
( `1234567` is revision of `based_on_tier.yaml` file, `bcd2345` is revision of `base/ns_dev.yaml`)
* `advanced-stage-1234567-cde3456`  
( `1234567` is revision of `based_on_tier.yaml` file, `cde3456` is revision of `base/ns_stage.yaml`)


for `base` it's simpler as it duplicates the same revision - it doesn't have the `based_on_tier.yaml` file
* `base-clusterresources-abc1234-abc1234`  
* `base-dev-bcd2345-bcd2345`   
* `base-stage-cde3456-cde3456`