feat: ChangeTierRequest updates Space
Apply the ChangeTierRequest on a Space with the same
name as the target MasterUserRecord, if it exists.
Only care about the `tierName` part of the ChangeTierRequest,
since the Space resource already caries the `targetCluster` info.

Also, rename all occurrences of `space` var to `s` to avoid conflict
with the `github.com/codeready-toolchain/host-operator/controllers/space`
pkg.

Fixes https://issues.redhat.com/browse/CRT-1311

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>