Allow user signups to complete when the associated space is updating (#876)
* Allow user signups to complete when the associated space is updating

If we mark user signups as incomplete while a space is updating, then
the registration controller will reject the login.  As a workaround,
allow user signups to progress while the associated space is updating.

Resolves https://issues.redhat.com/browse/SANDBOX-72.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* only allow completed signups in updating spaces

If a space is updating, then any associated user signups should be
allowed if and only if they've already been marked as complete.
Otherwise, wait until the space transitions from updating to ready
before processing signups.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* apply demorgan's laws

should be easier to read

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* refactor tests

The tests here had a dependency between executions.  Refactor them to
remove this dependency and reduce duplication between tests.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>
Co-authored-by: Matous Jobanek <mjobanek@redhat.com>