Implement a clean-up controller for TierTemplates.
This improves the system in two ways:

1. The TierTemplates now have a finalizer that prevents them from being
   deleted as long as any NSTemplateTier uses them.
2. Any unused TierTemplates are automatically deleted.

The first change is necessary so that the tiertemplates don't disappear
before all Spaces and NSTemplateSets that reference them are cleared up
first.

The second change is there to automatically clean up the tier templates 
as soon as the associated NSTemplateTiers are removed. I.e. we should
no longer leave "garbage" behind when removing NSTemplateTiers
(which is a rare occurrence but happens from time to time).